requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
gunicorn==21.2.0
